pyproject.toml
[tool.poetry]
name = "duck-duck"
version = "1.8.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]
readme = "README.md"
packages = [{include = "duck_duck"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
django-import-export = "^3.2.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.0"
factory-boy = "^3.3.0"
django-silk = "^5.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
